refactor: separate app metadata from macOS UI
